+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 ## [Unreleased]
 ### [0.15.1-alpha4] - 2025-01-20
 
+- performance improvement
+
 ## [0.15.1-alpha3] - 2025-01-20
 
 - fix doc mistakes