+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 ### [0.15.1-alpha4] - 2025-01-20
 
 - performance improvement
+- fix bug in feature `sprite`
 
 ## [0.15.1-alpha3] - 2025-01-20