+++ CHANGELOG.md
@@ -9,11 +9,12 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 -->
 
 ## [Unreleased]
-### [0.15.1-alpha4] - 2025-01-20
+## [0.15.1-alpha4] - 2025-01-20
 
 - performance improvement
 - fix bug in feature `sprite`
 - fix bug caused by duplicate added updating system
+- remove unnecessary type parameter in `QuadTree::query`
 
 ## [0.15.1-alpha3] - 2025-01-20